backend/requirements.txt
aiofiles==25.1.0
openai==1.30.0
fastapi-mail==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
Jinja2==3.1.3
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
greenlet==3.3.2
cryptography==46.0.7
ecdsa==0.19.2
et_xmlfile==2.0.0
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.24
PyYAML==6.0.3
redis==7.4.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
